Break out class wrapping
