+++ src/isolateComponent.ts
@@ -4,53 +4,16 @@ import { nodeTree, NodeTree } from './nodeTree'
 import { IsolateComponent } from './types/IsolateComponent'
 import { IsolatedComponent } from './types/IsolatedComponent'
 import { Selector } from './types/Selector'
+import { wrapClassComponent } from './wrapClassComponent'
 
 type Contexts = { contextType: React.Context<any>; contextValue: any }[]
+
 type RenderMethod<P> = (props: P) => any
 
 const getRenderMethod = <P>(t: any): RenderMethod<P> => {
   let proto = t.prototype
   if (proto?.isReactComponent) {
-    const instance = new t()
-
-    let first = true
-    let lastResult: any = null
-
-    return (props: P) => {
-      const [componentState, setComponentState] = useState(instance.state)
-
-      instance.setState = (s: any) => {
-        if (typeof s === 'function') throw new Error('not yet implemented')
-        const nextState = { ...componentState, ...s }
-        setComponentState(nextState)
-      }
-      const shouldRender =
-        !!instance.shouldComponentUpdate && !first
-          ? instance.shouldComponentUpdate(props, componentState)
-          : true
-
-      instance.props = props
-      instance.state = componentState
-
-      useEffect(() => {
-        if (instance.componentDidMount) {
-          instance.componentDidMount()
-        }
-        return () => {
-          if (instance.componentWillUnmount) {
-            instance.componentWillUnmount()
-          }
-        }
-      }, [])
-
-      if (shouldRender) {
-        lastResult = instance.render()
-        if (instance.componentDidUpdate && !first) instance.componentDidUpdate()
-      }
-
-      first = false
-      return lastResult
-    }
+    return wrapClassComponent(t)
   }
 
   return t
